Fix/framing fsm (#27)
* Updates.

* Updates.

* Updates.

* Updates.

* Updates.

* Updates.

+++ libraries/crc/src/crc.c
@@ -8,18 +8,18 @@
 
 static inline uint8_t _reverse_8bits(uint8_t byte)
 {
-    byte = ((byte & 0xF0) >> 4) | ((byte & 0x0F) << 4);  // Swap nibbles
-    byte = ((byte & 0xCC) >> 2) | ((byte & 0x33) << 2);  // Swap pairs
-    byte = ((byte & 0xAA) >> 1) | ((byte & 0x55) << 1);  // Swap bits
+    byte = (uint8_t)((byte & 0xF0) >> 4) | (uint8_t)((byte & 0x0F) << 4);  // Swap nibbles
+    byte = (uint8_t)((byte & 0xCC) >> 2) | (uint8_t)((byte & 0x33) << 2);  // Swap pairs
+    byte = (uint8_t)((byte & 0xAA) >> 1) | (uint8_t)((byte & 0x55) << 1);  // Swap bits
     return byte;
 }
 
 static inline uint16_t _reverse_16bits(uint16_t value)
 {
-    value = ((value & 0xFF00) >> 8) | ((value & 0x00FF) << 8);  // Swap bytes
-    value = ((value & 0xF0F0) >> 4) | ((value & 0x0F0F) << 4);  // Swap nibbles
-    value = ((value & 0xCCCC) >> 2) | ((value & 0x3333) << 2);  // Swap pairs
-    value = ((value & 0xAAAA) >> 1) | ((value & 0x5555) << 1);  // Swap bits
+    value = (uint16_t)((value & 0xFF00) >> 8) | (uint16_t)((value & 0x00FF) << 8);  // Swap bytes
+    value = (uint16_t)((value & 0xF0F0) >> 4) | (uint16_t)((value & 0x0F0F) << 4);  // Swap nibbles
+    value = (uint16_t)((value & 0xCCCC) >> 2) | (uint16_t)((value & 0x3333) << 2);  // Swap pairs
+    value = (uint16_t)((value & 0xAAAA) >> 1) | (uint16_t)((value & 0x5555) << 1);  // Swap bits
     return value;
 }
 
@@ -52,7 +52,7 @@ int8_t crc8_calculate(const struct crc* crc, const uint8_t* data, size_t length,
         {
             if (crc_value & 0x80)
             {
-                crc_value = (crc_value << 1) ^ crc->crc8_polynomial;
+                crc_value = (uint8_t)((crc_value << 1) ^ crc->crc8_polynomial);
             }
             else
             {
